Move get_libname() to common
This function, originally created for nut-scanner, to be able to lookup for
libraries loaded using libtool loader, is now also useful for the DMF support.
This last loads libneon at runtime, to load DMF structure (including the one
for nut-scanner) and unload it once the file were read
