+++ tools/nut-scanner/nutscan-init.c
@@ -24,7 +24,6 @@
 #include "common.h"
 #include <ltdl.h>
 #include <unistd.h>
-#include <dirent.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -41,48 +40,6 @@ int nutscan_load_neon_library(const char *libname_path);
 int nutscan_load_avahi_library(const char *libname_path);
 int nutscan_load_ipmi_library(const char *libname_path);
 int nutscan_load_upsclient_library(const char *libname_path);
-
-/* FIXME: would be good to get more from /etc/ld.so.conf[.d] */
-char * search_paths[] = {
-	LIBDIR,
-	"/usr"LIBDIR,
-	"/usr/lib64",
-	"/lib64",
-	"/usr/lib",
-	"/lib",
-	"/usr/local/lib",
-	NULL
-};
-
-const char * get_libname(const char* base_libname)
-{
-	DIR *dp;
-	struct dirent *dirp;
-	int index = 0;
-	char *libname_path = NULL;
-	char current_test_path[LARGEBUF];
-
-	for(index = 0 ; (search_paths[index] != NULL) && (libname_path == NULL) ; index++)
-	{
-		memset(current_test_path, 0, LARGEBUF);
-
-		if ((dp = opendir(search_paths[index])) == NULL)
-			continue;
-
-		while ((dirp = readdir(dp)) != NULL)
-		{
-			if(!strncmp(dirp->d_name, base_libname, strlen(base_libname))) {
-				snprintf(current_test_path, LARGEBUF, "%s/%s", search_paths[index], dirp->d_name);
-				libname_path = realpath(current_test_path, NULL);
-				if (libname_path != NULL)
-					break;
-			}
-		}
-		closedir(dp);
-	}
-	/* fprintf(stderr,"Looking for lib %s, found %s\n", base_libname, (libname_path!=NULL)?libname_path:"NULL");*/
-	return libname_path;
-}
 
 void nutscan_init(void)
 {
